Add license headers, move Never, restore For
