+++ src/main/java/io/github/fabriccommunity/everything/api/event/v4/events/OpenMenuEvent.java
@@ -1,3 +1,20 @@
+/*
+ *     Copyright (C) 2020 Fabric Community
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU Lesser General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU Lesser General Public License for more details.
+ *
+ *     You should have received a copy of the GNU Lesser General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 package io.github.fabriccommunity.everything.api.event.v4.events;
 
 import io.github.fabriccommunity.everything.api.event.v4.AbstractVetoableEvent;
